requirements.txt
Flask==2.2.3
firebase-admin>=6.2.0
requests>=2.31.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2
pycryptodome>=3.19.0
python-dotenv>=1.0.0
tqdm>=4.66.1 